01-rag-policy-advisor/requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
chromadb
pdfplumber
datasets
langfuse
